+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changes to Balzac.js
 
+### 0.0.2 (June 27, 2023)
+
+* Add Date.DURATION object to help converting delays into milliseconds.
+* Rename RegExp.INPUT_PATTERN object to PATTERN.
+* Fix RegExp.PATTERN.EMAIL, RegExp.PATTERN.EMAILS and RegExp.PATTERN.PHONE expressions.
+* Add RegExp.PATTERN.IBAN expression and related "iban" HTML input element.
+* Add RegExp.PATTERN.POSTCODE expression and related "postcode" HTML input element.
+* Remove whitespaces from both ends of text input values when case="upper" or case="lower".
+* Rename HTMLInputElement.PASSWORD_RATING.FAIR constant to MEDIUM.
+* Update README.md and test.html files.
+
 ### 0.0.1 (June 25, 2023)
 
 * Initial release